refactor: split main.go into cmd/ package (#111, #112, #113)
Split the 1228-line monolithic main.go into a cmd/ package with
per-command files:

- cmd/root.go: root command, Execute(), Version var
- cmd/helpers.go: path resolution, parseByteSize, humanSize, openInBrowser
- cmd/run.go: runFlags, addFilterFlags, buildFilter, runCmd
- cmd/list.go, configs.go, version.go, validate.go, check_env.go,
  trends.go, rerender.go, serve.go, plugins.go, clean.go, new_prompt.go
- cmd/cmd_test.go: all existing tests (moved from main_test.go)

main.go is now a 13-line thin entry point calling cmd.Execute().

No behavior changes — purely mechanical code movement.
