+++ hyoka/cmd/version.go
@@ -1,0 +1,17 @@
+package cmd
+
+import (
+"fmt"
+
+"github.com/spf13/cobra"
+)
+
+func versionCmd() *cobra.Command {
+return &cobra.Command{
+Use:   "version",
+Short: "Print version",
+Run: func(cmd *cobra.Command, args []string) {
+fmt.Printf("hyoka version %s\n", Version)
+},
+}
+}
